+++ unified_ai_client/__init__.py
@@ -15,6 +15,8 @@ from unified_ai_client.exceptions import (
     UnsupportedFileError,
     MissingFileError,
     FileDecodeError,
+    ProviderHttpError,
+    NonRetryableHttpError,
 )
 from unified_ai_client.silence import silence_sdks
 from unified_ai_client.verbosity import set_verbosity
@@ -38,6 +40,8 @@ __all__ = [
     "UnsupportedFileError",
     "MissingFileError",
     "FileDecodeError",
+    "ProviderHttpError",
+    "NonRetryableHttpError",
     "AiResponse",
     "AiRequest",
     "ProviderConfig",